refactor: reorganize libs
