+++ packages/plugin-views/src/lib/__tests__/configure.test.tsx
